Fixed warning, check if message is null, display error
Added a check to see if the message is null and if so display an error back to the user. Supressed the possible null warning later down, since it wouldent get to that point if it was null.

Warning:
src\Chirp.Web\Pages\SubmitCheep.cshtml.cs(21,13): warning CS8604: Possible null reference argument
 for parameter 'message' in 'CheepDTO.CheepDTO(string author, string message, long unixTimestamp)'